It works with an example file
The parser get simplified peptides in which only the target
modification is present. That will allow to look for the peptide in the
sequences and incorporate the modification in the proper location
according to recent databases.

The script is still not able to deal with different types of
modifications, different types of peptide encodings and neither is
outputing the conditional data.

Much more stuff to do