The residue positions are double checked when possible
There are 2 different cases. When we have the number of ptms in a
peptide is equivalent to the number of entries for this peptide. In
this case the mapping is exact and the mapping is perfect. For those
cases we check if the residue announced as a single column (if
available) is the same as the one we are detecting on the peptide. If
the column is not available we use the one in the peptide.

If the ptms in the peptide and the number of entries for this peptide
is not equivalent no validation is possible and therefore the residues
reported are ones found in the peptide